Add files via upload
DockUpdater.ps1 will detect a supported HP dock or docking monitor and obtain the installed firmware package version. It can also update the firmware to a newer version, if available. The script will use HP CMSL, if available in the client, to find the latest firmware for the attached device from the Internet, or use the version of the firmware from the hardcoded Softpaq links - mod as need for future needs